fix!(JuegoMesa): Corregir tipo de dato de llave

+++ backend/src/main/java/com/debede/backend/repository/Juego_De_MesaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface Juego_De_MesaRepository extends JpaRepository<Juego_De_Mesa, Long> {
+public interface Juego_De_MesaRepository extends JpaRepository<Juego_De_Mesa, Integer> {
 
 }
